+++ dashboard-widgets.el
@@ -160,7 +160,8 @@ preserved."
 The format is: 'icon title help action face prefix suffix'.
 
 Example:
-'((\"☆\" \"Star\" \"Show stars\" (lambda (&rest _) (show-stars)) 'warning \"[\" \"]\"))"
+'((\"☆\" \"Star\" \"Show stars\" (lambda (&rest _)
+                                    (show-stars)) 'warning \"[\" \"]\"))"
   :type '(repeat (repeat (list string string string function symbol string string)))
   :group 'dashboard)
 
@@ -1076,7 +1077,7 @@ It is the MATCH attribute for `org-map-entries'"
 When the dashboard-agenda is created this format is inserted into
 `org-agenda-prefix-format' as `dashboard-agenda' and compiled with
 `org-compile-prefix-format' previous calling `dashboard-agenda-entry-format' for
- each agenda entry."
+each agenda entry."
   :type 'string
   :group 'dashboard)